Skill tree additions and sound settings. Minor game stat changes.
Lowered customer payments and increased msg bonus to make more incentive to kidnap Customers. Added adjustable sound settings. Added increased wait time skill. Added increased distraction time skill. Bluetooth animatronic still in the works.

+++ Assets/Scenes/Main Folder/Scripts/Foodie/Foodie.cs
@@ -26,7 +26,7 @@ public class Foodie : MonoBehaviour
     
 
     [Header("-----ORDERING SETTINGS-----")]
-    public int orderTime = 28;
+    public float orderTime = 28;
 
     public float timeAtOrderTaken;
     public Sprite order;
